Added Dragon Crafting Recipes
Added crafting recipes and config file options relating to the Ender
Dragon, such as for End Crystals, Dragon Egg, and Dragon's Breath. Also
updated the comment for the Alternate Carrot on a Stick Crafting Recipe
to match the recipe and updated the horse armor log output to change
depending on whether or not saddles are used.